[IMP] knowledge: allow embedded views to stick to the top
This commit adds a new css rule to enable articles starting by an embedded view
to stick to the top of the article.

This display enables the user to create articles with only their embedded view
inside it.

This commit also removes the insertion of titles inside the article's body for
the helper's templates.
These templates should better represent an article containing only an embedded
List/Kanban/Calendar.

It also adds a new plugin to Knowledge that handles the deletion of an article's
1st line. This way when a user wants a clean article with only the embedded view
on top it's easier to do so. This plugins also ensures that every article
contains at least a paragraph if it happens the user deletes one too many.

task-3231001

closes odoo/enterprise#38529

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>