[IMP] project_enterprise: add critical path in project gantt view
Purpose
=======
The critical path shows those tasks that must be completed on
time. It is the longest chain of tasks between the start and end dates
of your project that cannot be compromised. It helps to know what
sequence of tasks requires special attention, and how much it takes to
complete your project on time.

Implementation
==============
The critical path is the longest path of dependent tasks of a given
project. This commit introduces:
- Backend methods to get the critical path of a group of tasks (based on
a domain)
- Frontend methods to display the critical path of a project in the
Gantt view
- Testing for those methods

task-3391988

closes odoo/enterprise#43440

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>