[IMP] timer: add timer parent mixin to uniform the using of timer
Before this commit, when the timer used in task, the user cannot
stopped it in the timesheet app and so the user has to keep the task
form view to be able to stop the timer.

This commit creates a timer parent mixin to be able to launch the timer
in one place. It means that newest mixin allows to add a link in the
timer model to another document. The idea with this mixin is to linked
the timer to the main document (that is timesheet model) and also to a
parent of the related document (that is the ticket or the task for
instance). Thanks to that, the user will be able to see the timer in the
parent document but also in the related document.

task-2276015