[IMP] timesheet_grid: align timer on tasks with timer header
Before this commit, when the timer is started in the task form view, it
was impossible for the user to stop the timer in the Timesheet app.

This commit uses the `timer.parent.mixin` in the `project.task` instead
of `timer.mixin` to be able to start/stop the timer in the timesheet app
and in the task form view. It means when the user will start the timer in
the task form view, a timesheet will be generated and the timer will be
linked to that timesheet. And so, the user will be able to stop the
timer in the timeshet app.

task-2276015