[IMP] helpdesk_timesheet: align timer on tickets with timer header
Before this commit, when the timer is started in the ticket form view, it
was impossible for the user to stop the timer in the Timesheet app.

This commit uses the `timer.parent.mixin` in the `helpdesk.ticket` instead
of `timer.mixin` to be able to start/stop the timer in the timesheet app
and in the ticket form view. It means when the user will start the timer in
the ticket form view, a timesheet will be generated and the timer will be
linked to that timesheet. And so, the user will be able to stop the
timer in the timeshet app.

task-2276015