[REM] helpdesk_timesheet: remove helpdesk.ticket.create.timesheet wizard
Before this commit, when the user stopped the timer in the ticket form
view, a wizard was displayed to generate the timesheet according to the
time elapsed with the timer. Since the timeshet is now generated when the
timer is started, this wizard will just update the existing timesheet.
This update could be done in the sub tree containing the timesheets
related to the ticket.

That's why, this commit removes this wizard.

task-2276015