[IMP] timesheet_grid,*: add confirmation dialog when timer stopped in task
* = helpdesk_timesheet, industry_fsm, industry_fsm_report

Before this commit, when the user stops the timer in project.task or
helpdesk.ticket form view he cannot easily edit the description of his
timesheet generated by the timer as it is the case in the Timesheets app.

This commit adds a confirmation dialog when the user stops the timer
to let him editing the time spent and the description of his timesheet
generated with the timer.

task-2276015