[IMP] l10n_dk_rsu: Send tax report
The Danish government has set up a system enabling users of accounting software
 to send their tax reports to the government directly from the accounting software.

To do this, there are 3 different web services:
- VirksomhedKalenderHent: This web service returns the dates on which the legal
 entity must submit its VAT returns. These dates are required to submit VAT returns.
- ModtagMomsangivelseForeloebig: This web service submits a draft VAT return to
 skat.dk with all the fields to be filled in. The web service returns a deep
 link to skat.dk, where the legal entity can access and approve the submitted
 VAT declaration.
- MomsangivelseKvitteringHent: This web service provides a receipt for the VAT
return, given that the legal entity has approved it. This service also includes
 information on how to pay any outstanding balance.

These three web services use SOAP and encryption to function. What we decided
to do was to take the SOAP envelopes of the requests and create them by hand
and send them to the government portals.

closes odoo/enterprise#92589

Task: 3326971
X-original-commit: 184a19a5763c4e73d5e264860888d8ce113b6d61
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>