[IMP] planning: copy out-of-schedule shifts as open
Before this commit, when a shift was out of its resource's schedule,
we identified what "half day" it was in and if the resource was
working this half day.
If so, we would assign it, else it would be copied as open.
So say a resource that works from monday to friday,
8-12 and 13-17. If, on the previous week, he had a shift on tuesday
(19-20), as he worked this afternoon, he would be assigned to the shift.

However, we now have "flexible resources" -technically, their field
`calendar_id` is set to False-, which are always available.
As a consequence, we can look at the calendar of the resources that
have one more literally. So, in this commit, we stop checking half days.
If the resource's schedule says he doesn't work at that time,
the shift is now copied as open.

Also, if a flexible resource is assigned to a shift,
we copy it no matter what and assign him.
Note that flexible resources have no schedule and
that it should be impossible for them to take time-off,
so none of the standard usecases where the shifts are copied as open
actually apply to them.

task-3231732

closes odoo/enterprise#44690

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>