[ADD] whatsapp_calendar: allow to send reminder using whatsapp
PURPOSE

The purpose of this commit is to introduce a new reminder of WhatsApp type for
calendar events. Now the user can send WhatsApp messages as a reminder of
calendar events by defining a WhatsApp template for the remainder, similar to
the existing functionalities for emails and SMS.

SPECIFICATION

A new feature is added by this module to send a reminder of calendar events
through WhatsApp. The users can create a reminder of WhatsApp type by defining a
WhatsApp template that applies to the `calendar.attendee` model. Users can also
define whether to send the remainder to the organizer of the calendar event or
not by defining the `notify responsible` boolean.

Related Upgrade PR: https://github.com/odoo/upgrade/pull/5709

Task-3506616

closes odoo/enterprise#56894

Related: odoo/odoo#154535
Related: odoo/upgrade#5709
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>