[FIX] l10n_de_reports: date scope income/loss for the year
Profit and loss from previous year was not well calculated
on the balance sheet.
With this commit, we change the date scope for the `Net profit/loss for the year`
line from `normal` to `from_fiscalyear`.
We also change the way we calculate `Profit carried forward/loss carried
forward` and `Net income/loss for the year` lines.

opw-3663626

closes odoo/enterprise#82448

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>