[FIX] web_grid,timesheet_grid: resolve missing content in action helper
Versions:
17.0

Steps to reproduce:
go to timesheet module
click on grid cell

Issue:
The action helper in web grid was resulting in only the web content being
displayed in the timesheet grid and so helper was missing

Cause:
The action helper was being called from web_grid, causing a lack of content in
the timesheet_grid.

Solution:
Create a dedicated function to retrieve the no-content helper, enhancing
portability for easy customization.

task-3429403

closes odoo/enterprise#81091

X-original-commit: 1f1a93f919b2855b639a16516ee809e378c0435d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>