[FIX] industry_fsm: fix SOI not set by default issue
- Install the Field Services (industry_fsm) and sales module.
- Create a sale order with the following configuration:
  - Select any customer.
  - Select the product 'Junior Architect' (Invoice on Timesheets).
  - Confirm the sale order.
- On the header of the sale order, click on the project option.
- Then, in the left side corner, click on 'Share Editable'.
- Share the project to your company.
- Log in as a portal user.
- Open the projects menu.
- Open the project you just shared.
- In the kanban view, click on the 'Quick Create Task' button.
- Add a task name and click on the edit button.
- You will see the form view where the customer and sale order item are not set
  by default.

Issue:
------------------

When logging in as a portal user(project sharing), the fields for Customer and
Sale Order Item  are not set by default when creating a task via quick create in
the Kanban view.

Cause:
-------------------
The  _compute_partner_id function relies on the project_id field to
determine the partner_id (customer) for a task. However, in the quick create
view, only the partner_id field is mentioned . in the quick view, the project
ID is not provided so the partner ID cannot be computed, resulting in an empty
field for the partner_id (customer).

Solution:
----------------------
add a project_id field in the 'project_sharing_quick_create_task_form_inherit'
quick create form view.

task-3743360

closes odoo/enterprise#71271

X-original-commit: d9b5b5f26e8f89cd78d5fc8d28c45d3c6d6be546
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>