[IMP] l10n_pe_edi: adding the XML structure to support free billing
We need to be able to invoice free items. How it works in Peru is that
the item is "sold" with the original price. Still, we should be able to
mark the transaction as a free transaction (this is done with the field
l10n_pe_edi_legend that is in the Peru EDI tab). This tag should change
the XML to the correct format for a free transaction

An enhancement is introduced to exclude the 'l10n_pe.tax_group_dummy' tax group from tax details processing.
This improvement is aimed at facilitating free billing in Peru by ensuring dummy tax groups are not included in tax calculations.

X-original-commit: 2c58315e843fbe4b4db525d0493bc963713e9b7a
Part-of: odoo/enterprise#74326
Related: odoo/odoo#188145
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>