[FIX] various: be defensive with activity types, define master data
See community PR for more details. In general, be defensive with
activity types: avoid crash when activity type has been removed,
protect types linked to business code that should not be unlinked
or changed from model.

Account Online Synchornization: make "Bank Synchronization" master
data as business behavior dependso on it. Also fix model used for
the type.

Account reports: make "Tax Closing", "Pay Tax" and "Tax Report Ready"
master data as business flow depends on it.

Approvals: make "Approval" master data as business flow depends on it
and it is not easy to remove it in their usage.

Hr Payroll: make "Leaves to defer" master data as business flow
depends on it.

Voip: make "Call" activity type master data users cannot remove as it
is required in various flows of VOIP. Also force its model to be False
as it is used in various models and should not suddenly be limited to
a given model. Done in community, as activity is defined in 'mail'.

Studio: make "Approval" master data as approval flow depends on it.

Task-3777606

X-original-commit: a45df29517b807d3868a26b1cd5030ad9c5b99b5
Part-of: odoo/enterprise#90235
Related: odoo/odoo#219005
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>