[IMP] account_*: Add a total of bank account with pending
Display the available balance for an account when possible. The available balance is tipicaly the balance plus/minus the amount of pending transactions. When available balance is different from balance, allow the user to fetch (but not import & reconcile) the pending transactions. This way the user knows what amount is actually available to him and he can check if payments due are already existing in pending state.

closes odoo/enterprise#59892

Task: 3638204
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>