[IMP] planning: make date changes impact recurrence more cleanly
If the user doesn't change any date field, the recurrence should be
impacted as expected.
However, when at least one date field is changed, with option:
- "all": the changes should be applied to the first shift, then
all other shifts in the recurrence should be deleted and recreated,
using the first one as template.
- "subsequent": the previous shift's recurrence should be unset, the
changes should be applied to current shift, the subsequent shifts should
be deleted and recreated, using the current one as template.

task-3765142

Part-of: odoo/enterprise#58302
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>