[FIX] planning: added filter groupby
Before this commit there was no groupby Properties.

In this commit we add a filter for groupby properties in planning module.
The search view is being used by two modules`planning.slot` and
`planning.analysis.report` thus solution resulted in adding the filter in
`planning_view_search` as it is a primary inherited view of search view base in
planning.slot.

task-3801551

closes odoo/enterprise#75794

X-original-commit: e2db2b85ffc2638cda3a1a236e8cd0d0c78df9f2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: K Theja (kthe) <kthe@odoo.com>