[FIX] pos_preparation_display: fix tracebackback when stages deleted in backend
This traceback arises when the user tries to change the orders from one stage to
another.

To reproduce this issue:-

1) Install "pos_restaurant"
2) Duplicate the tab and open preparation display in one tab[1]
3) In another tab delete the last stage(complete) for the above preparation display
4) In first tab click any product of an order from the "ready" stage
5) Now click the ready icon of that order
5) At terminal you will see a traceback

Error:- "IndexError: list index out of range"

When the user click on current stage icion it supposed to go to next stage
with the help of current stage index + 1.

But, because the user deleted the next stage it leads to index error

After applying this will resolve the issue by preventing the user to modify stages
when there is an active session for that preparation display. Which making the code more robust

sentry-5056416790

closes odoo/enterprise#80090

X-original-commit: f980f220602cb564535201cfdcf73b5b2ba6046e
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>