[IMP] project_forecast,_*: multi-company count of stat buttons in project updates
Before this commit, the stat buttons count values in the project updates
right-side panel were not updated when the user was changing companies
in the company selector.
After this commit, the counts displayed in the stat buttons are updated
based on the selected companies.

task-3689065
related-https://github.com/odoo/odoo/pull/150870

closes odoo/enterprise#55033

Related: odoo/odoo#150870
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>