[IMP] industry_fsm_appointment, website_appointment_sale_project: better integration with appointment
This commit introduce a bridge module for Field Service, Appointment and Websale,
allowing to set data on the task of Field Service or Project, according to
the data defined in the appointment. It also adds a stat button to navigate
to the appointment from the task view.

Prior to this commit, when a SO is confirmed through the Appointment Web
Interface, for an appointment type which creates a Task in Field Service
or Project, a task is created but with two issues:
1. There is no link between the Appointment and the related task.
2. Following data are missing in the task: selected users or resources,
   duration, questions and answers of the appointment.

The following fields are set on the task:
- Description includes Q&A of the appointment
- User assigned to the appointment is assigned to the task
- Guests are set as followers of the task
- Duration is set according to the appointment
- the invited guests of the appointment are now added as followers of the created task.
- Some formatting applied to the Q&A of the appointment, shown in the descrpition of the task.
- Tests have been modified to verify the followers as well.

Task-3823058

closes odoo/enterprise#66576

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>