[FIX] appointment: set condition based on type
Added condition on a check in microsoft calendar that ensured the organizer
was an attendee on the event on create. This was added in a change that allowed the organizer to be
changed on the Odoo side. However, this caused issues with the
appointments app when creating an appointment that only used resources.
The "organizer" who in this case is the creator of the appointment type
would not be an attendee and therefore would cause an error and make the
appointment type unbookable.

Adding this condition allows the appointment to bypass the check if the
appointment type uses resources instead of users.

This change adds the necessary override in enterprise. See change in
community here: https://github.com/odoo/odoo/pull/178142

foward-port of 54edf00c6ad81f8143983820ecb43501199998fe

opw-3841495

X-original-commit: 38400ccacd40701909b3192cddd1e5b38774ffb3
Part-of: odoo/enterprise#83168
Related: odoo/odoo#205398
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>