[IMP] project_timesheet_holidays: improve Generate Timesheets Feature
Before this PR:
By default, the current company is set on time off types. As such, the internal
project of the current company is set.

After this PR:
This commit improves the functionality related to generating timesheets within
the project_timesheet_holidays module. Specifically,  it removes the association
of projects and tasks from the time-off type and hides generated timesheets.
Removed timesheet generate field, as it is not used in any inherited view.

Reason:
 if we set the company of the time off type to false, we either have:
- no project -> meaning the user will disable the feature without realizing as
the field is only visible in debug mode;
- the internal project of the current company that remains set ->
 meaning the feature won't work for the other companies in which a time off
 request is created.

task- 3605962

closes odoo/enterprise#62577

Related: odoo/odoo#165674
Related: odoo/upgrade#6027
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>