[FIX] sign: signature sign submit buttons disabled when no sign
Issue: After the changes we have made to sign in 17.0 we stop properly
handling the state of the buttons for submit our sign after signing,
the problem then is that in Draw mode for example, we are able to submit
the sign even before drawing anything, which will allow us to send the
document "unsigned", since there will not be any sign.

Steps to reproduce:

1. Get Sign module.
2. Upload any document to sign.
3. Add sign box and go to sign the document.
4. Now change from Auto to Draw.

Solution: We should take into account that it makes sense that everytime
we update if our signature is empty, we update the state of the buttons
accordingly avoiding this way to be able to sign without an actual sing.
The solution address a similar approach at what was already being done
with the onChangeName, we move this logic to be handled by a specific
function for this buttons 'handleButtonStateChange'.

opw-3874034

closes odoo/enterprise#74071

X-original-commit: e9fb7c4b50a34d6ab70342a904830a8c8af54c8b
Related: odoo/odoo#187673
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>