[IMP] worksheet: allow user who access to task to access to worksheet
Before this commit, if a worksheet is created by a user A on a task and
another fsm user goes to that task and try to access to the worksheet,
that other user will not have access to the worksheet if the user is not
at least a fsm user or a project manager, he will just has access to his
own worksheet if he is a project user.

This commit adds a new ir.rule for the project user to give access to the
worksheet linked to a task in which the users are assigned.

task-3891011

closes odoo/enterprise#62542

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>