[FIX] planning : Take resource and working intervals into consideration
Steps to reproduce:
	- Create a working calendar 8:00 - 12:00 13:00 - 15:00
	- Set in employee and company, as you mentioned everywhere.
	- The create a shift template 11 :00 - 15:00 (3 hrs)
	- Now create a planning slot with this shift template.
	- Expected start date - End date 11:00 - 15:00
	- Actual start - end 11:00 - 14:00

Current behavior before PR:
The end time is not calculated correctly if we have working calendar
because we just add the duration to the start time without taking
into consideration that there might be a break hour in the middle of
the duration.

Desired behavior after PR is merged:
We are now considering the employee resource calendar in
our calculation for the time of the slot.

Also changed 'test_calculate_slot_duration_flexible_hours' as we are now
calculating allocated hours differently.
opw-3890629

closes odoo/enterprise#72321

X-original-commit: 959dd7e11fa3a2c2892ecb28a06f0bbd76965772
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>