[IMP] sign: improve assignment of sign requests to contacts
Before this commit:
When creating two contacts with a sales email, one linked to a user and the
other not, the sign request would always be assigned to the contact with the
smallest ID, disregarding user association.

After this commit:
The assignment of sign requests now checks if the contact (partner) is linked to
a user. If a user is associated, the request will be assigned accordingly.

task-3971110

closes odoo/enterprise#64200

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>