[FIX] l10n_in_gstr_reports: remove CGST/SGST from overseas transactions
Before this commit, overseas transactions with GST were shown in the
spreadsheet, but not in GSTR-1.

After this commit, overseas transactions with GST will no longer be shown in the
spreadsheet.

Technical Reason:-
In the `_get_exp_json` method, we retrieve journal items that meet the exp
domain criteria. However, the `tax_details` are linked from `move_id` using
`tax_details_by_move`, which applies the hsn domain. Consequently, all journal
items are retrieved, necessitating the exclusion of those involving CGST and
SGST, as these are not relevant for overseas transactions.

Task-4008149

closes odoo/enterprise#71631

X-original-commit: 29a12efc81ee584d6be72fbb0618d7adc99b078e
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Jay Savaliya (jasa) <jasa@odoo.com>