[IMP] social_twitter: apply new private likes logic
The X platform announced that they were updating their policy on likes. These
changes also impacts their API, this commit updates the social_twitter module to
handle this change.

The API has been changed so that when accessing the Likes lookup endpoints, the
resulting response will contain data about the accessed posts if it belongs to
the authenticated user.

It is still possible to fetch the like_counts for all posts, but not the details
of each like.
This disables any user to evaluate who is liking its posts.
This also blocks any attempt to retrieve the posts liked by a specific user.

If a user tries to access these endpoints the resulting response will be:
```
{
    "meta": {
        "result_count": 0
    }
}
```

These responses aren't considered errors so exisiting streams using this
endpoint will not crash.
But they will not contain any fresh post and won't be able to be refreshed.

task-3986158

Part-of: odoo/enterprise#64647
Related: odoo/upgrade#6158
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>