[IMP] sign: use the same date format when auto-filling date fields
Before this commit, it was possible to fill two date
fields by two different date formats in the same document.
This happend because the date format was
dependent on the location of user who is signing.

After this commit, date fields will be auto-filled by a fixed
date format, that format depends on the language of the company's partner.

closes odoo/enterprise#72363

Task: 3930358
X-original-commit: ca5c0368c98d1d32f1a7cac2d8ee4a1ae0f6711c
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>