[FIX] whatsapp: avoid double-sending the same body
When sending templates en masse to multiple recipients
they sometimes may have the same number.

If their numbers are the same, and the template resolves to the same value
it is useless to send them the message multiple times.
This behavior could be viewed as spam.

Now if all template variables resolve to the same value, messages will be
cancelled immediately. Similarly to what is done for emails.

task-4029026

X-original-commit: 4e7cdefc301c72a9016f0a4b57835536466c6c44
Part-of: odoo/enterprise#78655
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>