[FIX] documents: information of message sender from alias
Before this commit:
The information of message sender using email alias was not shown,
instead only mail subject with name Odoobot was displayed in chatter.

After this commit:
The information of sender such as email address, name, subject of email
and email body is not displayed in chatter along with the attachment.

Task - 4008634

closes odoo/enterprise#71885

X-original-commit: 52a99f5b7c29ff9ed1825d415b8034d3c0808035
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>