[REM] l10n_au_reports: conditional import of a non-function
As far as I can tell `stdnum.au.abn.is_valid_abn` does not exist and
has never existed. There is an `stdnum.au.abn.is_valid` which has been
present since the module was first added in stdnum 1.5 but it's not
clear if that's the one we want here.

So in doubt maintain functional equivalence by doing nothing.

Part-of: odoo/enterprise#66490
Related: odoo/odoo#172761
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>