[FIX] marketing_automation: do not commit in test mode
We should always avoid hard "commits" in test mode, as it is quite
incompatible with rollbacks and the way test classes handle the cursor.

Synchronizing participants in marketing automation create records
in batch, and commit after batches in non-test mode. However it also
unlinks some records and that part is always committed.

In this commit we correctly commit when not in test mode for all parts
of the synchronizing process.

Task-4605118: [marketing_automation] Unbordelize test classes
Prepares Task-4224152: [marketing_automation] Performance / Scalability

Part-of: odoo/enterprise#80180
Related: odoo/odoo#199336
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>