[FIX] account_reports_cash_basis: make sure the domain is applied on the correct model
Issue
----

When clicking an account report cell with the cash basis filter on,
`account_reports_cash_basis.account_report.action_audit_cell` adds a domain to filter out
`account.move.line`s that don't belong to the cash basis criteria. However, the model is
not always an `account.move.line`. If an analytic account column is added with the
"Analytic Group By" option, the model is `account.analytic.line`, which results in a field
not found error.

Steps
----

 - Go to Accounting -> Reporting -> Profit and Loss.
 - From the Options button above, select "Cash Basis Method".
 - From the Analytic Group By button, add an analytic account in the "Accounts" filter.
 - Click on a cell under the analytic account's column.
 - An error is thrown.

Cause
----

Domain is applied on the wrong model.

opw-4034786

closes odoo/enterprise#70350

X-original-commit: 2a17180467ff6a6830c34b30a6ce2dc3ef94649d
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>