[ADD] l10n_cl_edi_factoring:
The l10n_cl_edi_factoring module allows the credit and rights of a generated sales invoice to be yielded to a third party. The third-party will advance the invoice money, charging a commission or interest on the amount as compensation. This feature enhances financial flexibility and cash flow management for businesses using our Odoo instance. The module complies with requirements from the SII (Servicio de Impuestos Internos) regarding the electronic execution of the yield.
It includes the generation of the AEC (Archivo Electrónico de Cesión) file, the creation of the account move for the yield, and the reconciliation of the invoice with the yield account move. This is also compatible with l10n_cl_edi demo mode, generating and simulating the send and acceptation of the AEC to the SII.

closes odoo/enterprise#73520

X-original-commit: 3e25b54286e1ce9852b43320989c82cafc308c91
Related: odoo/odoo#186679
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>