[FIX] account_batch_payment: Prevent adding batched/validated payments
Issue:
When we create a batch from a payment by selecting it and clicking on
"Create Batch", even if the payment is already in a batch or validated,
it is possible to add it to a new batch, which can lead to issues when
validating the batch.

Steps to reproduce:

- Create a payment (posted but not sent) SEPA transfer.
- Select this payment and create batch payment.
- validate the batch
- Open payments list again, and select that same payment, then create
a batch payment from the selection.
- a new batch payment will be created.
- If you open the batch payments list you'll notice 2 batches:
1-the original batch now has no payments inside it so the "re-generate
Export File" button won't show up.
2- the second batch now contains the payment, but we can't re-generate
the export file because it shows an error:
The batch could not be validated, Some payments have already been sent.

Solution:
- before creating a batch with payments or adding payments to a batch
we check if the payments we are adding are already in a batch.

OPW-3976563

closes odoo/enterprise#75898

X-original-commit: 63fee72313416a6393cfe557834cd71341c034f3
Signed-off-by: Julien Van Roy (juvr) <JulienVR@users.noreply.github.com>
Signed-off-by: Kawtar Drissi El Bouzaidi (kdeb) <kdeb@odoo.com>