[IMP] industry_fsm: show project name if there is only one company
In the project list view, instead of displaying the name, we use the display_name so that
we can show the company name for fsm projects.
However, since it is a computed field, it has the drawback of not being sortable.

To alleviate this, this commit will use the name instead if there is only one company in the context
or if Field Service is not installed.

Task-3942760

closes odoo/enterprise#67791

Related: odoo/odoo#175324
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>